pyproject.toml
[project]
name = "ylz-utils"
version = "0.5.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.3",
    "fastapi>=0.115.12",
    "fastapi-cdn-host>=0.9.1",
    "sse-starlette>=2.3.3",
    "uvicorn>=0.34.2",
    "httpx>=0.28.1",
    "mpmath>=1.3.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "sqlit>=0.1.6",
    "aiosqlite>=0.21.0",
    "apscheduler>=3.11.0",
    "cryptography>=44.0.2",
    "huggingface-hub>=0.30.2",
    "modelscope>=1.25.0",
    "langchain>=0.3.24",
    "langchain-community==0.3.24",
    "langchain-huggingface>=0.1.2",
    "langchain-modelscope-integration>=0.1.0",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.2.14",
    "langchain-together>=0.2.0",
    "langchain-google-genai>=2.1.3",
    "dashscope>=1.23.2",
    "qianfan>=0.4.12.3",
    "langgraph>=0.3.34",
    "langgraph-sdk>=0.1.63",
    "grandalf>=0.8",
    "duckduckgo-search>=8.0.1",
    "addict>=2.4.0",
    "langchain-mcp-adapters>=0.0.9",
    "faiss-cpu>=1.11.0",
    "transformers",
    "accelerate>=1.6.0",
    "datasets==2.21.0",
    "langchain-experimental>=0.3.4",
    "langserve>=0.3.1",
    "librosa>=0.11.0",
    "soundfile>=0.13.1",
    "diffusers>=0.33.1",
    "qwen-omni-utils>=0.0.4",
    "torchvision>=0.22.0",
    "click>=8.1.8",
    "chromadb>=0.6.3",
    "psycopg[binary,pool]>=3.2.9",
    "langgraph-checkpoint-postgres>=2.0.21",
    "langchain-cloudflare>=0.1.8",
]

[project.scripts]
ylz_utils = "ylz_utils.cli.main:run"

[project.optional-dependencies]
tools = [
    "akshare>=1.16.83",
    "gradio>=5.27.1",
    "gradio-client>=1.9.1",
    "gradio-tools>=0.0.9",
    "ipfs-toolkit>=0.5.33",
    "matplotlib>=3.10.1",
    "playwright>=1.51.0",
    "serpapi>=0.1.5",
    "tabulate>=0.9.0",
    "wolframalpha>=5.1.3",
]
docs = [
    "elasticsearch-dsl>=8.18.0",
    "langchain-elasticsearch>=0.3.2",
    "neo4j>=5.28.1",
    "docx2txt>=0.9",
    "langchain-chroma>=0.2.3",
    "markdownify>=1.1.0",
    "pandas>=2.2.3",
    "pdfminer>=20191125",
    "pdfplumber>=0.11.6",
    "pillow>=11.2.1",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "trustcall>=0.0.39",
    "pypdf>=5.5.0",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers" }




